refactor(auth/profile): implement dependency injection and service interfaces
- Add DI container middleware and service interfaces
- Refactor controllers to use injected services

+++ src/services/profile.service.ts
@@ -2,34 +2,36 @@ import { eq } from "drizzle-orm";
 import { db } from "@/db";
 import { profiles } from "@/db/schema";
 import { IProfileInsert } from "@/types/profile.type";
+import { type IProfile } from "./../db/schema";
+import { IProfileService } from "@/model/profile/profile-services.interface";
 
-export class ProfileService {
-  async getProfile(userId: string) {
+export class ProfileService implements IProfileService {
+  async getProfile(userId: string): Promise<typeof IProfile | null> {
     const [profile] = await db
       .select()
       .from(profiles)
-          .where(eq(profiles.userId, userId));
-          
-    if (!profile) return null;
-      
-    return profile;
-    }
-    
-    async saveProfile(profileData: IProfileInsert, userId: string) {
-      await db.insert(profiles)
-      .values({
-        userId,
-        firstName: profileData.firstName,
-        lastName: profileData.lastName,
-        phoneNumber: profileData.phoneNumber,
-        address: profileData.address,
-        birthDate: profileData.birthDate,
-        gender: profileData.gender,
-      })
-    }
+      .where(eq(profiles.userId, userId));
 
-    async updateProfile(profileData: IProfileInsert, userId: string) {
-      await db.update(profiles)
+    if (!profile) return null;
+
+    return profile;
+  }
+
+  async saveProfile(profileData: IProfileInsert, userId: string) {
+    await db.insert(profiles).values({
+      userId,
+      firstName: profileData.firstName,
+      lastName: profileData.lastName,
+      phoneNumber: profileData.phoneNumber,
+      address: profileData.address,
+      birthDate: profileData.birthDate,
+      gender: profileData.gender,
+    });
+  }
+
+  async updateProfile(profileData: IProfileInsert, userId: string) {
+    await db
+      .update(profiles)
       .set({
         firstName: profileData.firstName,
         lastName: profileData.lastName,
@@ -39,10 +41,9 @@ export class ProfileService {
         gender: profileData.gender,
       })
       .where(eq(profiles.userId, userId));
-    }
+  }
 
-    async deleteProfile(userId: string) {
-      await db.delete(profiles)
-      .where(eq(profiles.userId, userId));
-    }
+  async deleteProfile(userId: string) {
+    await db.delete(profiles).where(eq(profiles.userId, userId));
+  }
 }
